chore: create a shared folder
