+++ assets/src/utils/common.js
@@ -80,41 +80,6 @@ const EDITOR_MAP_ARCHIVED = {
 	},
 };
 
-/**
- * Loop through an Elementor element and apply the function.
- *
- * @param {any} element  Elementor element.
- * @param {*} applyFunc The function to apply on each child element.
- */
-const loopElementorElement = ( element, applyFunc ) => {
-	applyFunc( element );
-
-	element?.elements?.forEach( ( item ) => {
-		loopElementorElement( item, applyFunc );
-	} );
-};
-
-/**
- * Clean the template content from unnecessary data.
- *
- * @param {any} templateContent The template content.
- * @param {Function} cleanFunc The function to apply on each element.
- * @return {any} The cleaned template content.
- */
-const cleanTemplateContent = ( templateContent, cleanFunc ) => {
-	if ( undefined === templateContent.content ) {
-		return templateContent;
-	}
-
-	const content = templateContent.content;
-
-	if ( Array.isArray( content ) ) {
-		content.forEach( ( item ) => {
-			loopElementorElement( item, cleanFunc );
-		} );
-	}
-};
-
 export {
 	addUrlHash,
 	getTabHash,
@@ -125,5 +90,4 @@ export {
 	EDITOR_MAP,
 	EDITOR_MAP_ARCHIVED,
 	TAGS,
-	cleanTemplateContent,
 };
